Update the implementations of socket.
Added socket to show the latitude and longitude of the next 300m
and mark it accordinly to mimic the real time location of a passing
ambulance.
The client side code still needs to be added. Currently an array
of the locations of simulated array is passed to the clients via
socket.io